Low: misc: docker autogen now attempts to use whatever IPs are assigned to containers
Previously we attempted to dynamically change the IP addresses of
the containers to what we wanted them to be.  Changes to docker
prevents us from using static ips using the method we were using.

Now we use whatever IP addresses docker assigns the containers instead
of fighting docker to assign the static address. In the future, the
hope is that we will be able to assign static IP addresses using
docker and everything will just work.

Since the ip addresses are not persistent, docker containers
will not beable to join the cluster if the container is
restarted... this really needs to change